test : lotto, lottos 테스트
